Move Element model to models/AXElement
